# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DocuForge API
  version: 1.0.0
  description: OpenAPI specification for DocuForge core services (auth, users, documents, conversions, security, admin)
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DocumentUploadResponse:
      type: object
      properties:
        document_id:
          type: string
        filename:
          type: string
        download_url:
          type: string
...
paths:
...
  /documents/upload:
    post:
      summary: Upload a document
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
...
  /documents/{id}/download:
    get:
      summary: Download document file
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File stream
...
  /security/unlock:
    post:
      summary: Unlock / decrypt a protected PDF (keep this feature)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                document_id:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Unlocked document created
...